changed the logic of ram cost.
Previously instance scores are kept same for each host, which actually result in a stacked solution for multi instance creation problems. Here the cost matrix is changed so that each instance get a different score according to the order in a host it is placed. Also it should be noticed that the meaning of variable matrix is also changed. Now host-instance matrix A is defined such that A[i][j] denotes the placement of a (j+1)'th instance in host i, and this doesn't correspond to a specific instance, which means that e.g. A[1][2] denotes whether a 3rd instance is placed in host 1, and if its value is 1 then the values of A[1][0] and A[1][1] must also be 1s because it is not possible to place a 3rd instance without placing 1st and 2nd instance.